app/portfolio_chat/pyproject.toml
[project]
name = "portfolio-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.10.11",
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.17",
    "langchain-pinecone>=0.2.6",
    "langgraph>=0.5.3",
    "pinecone>=6.0.2",
    "pydantic>=2.11.4",
    "python-telegram-bot>=22.3",
    "slowapi>=0.1.9",
    "uvicorn>=0.34.2",
]
